feat: major refactor - split handler, add DI, sessions, migrations, mocks, CI pipeline

+++ internal/domain/analytics.go
@@ -5,23 +5,23 @@ type AnalyticsInsight struct {
 	Title       string  `json:"title"`
 	Description string  `json:"description"`
 	Severity    string  `json:"severity"`
-	Value       float64 `json:"value,omitempty"`
+	Value  int64  `json:"value,omitempty"`
 	Metric      string  `json:"metric,omitempty"`
 	Period      string  `json:"period,omitempty"`
 }
 
 type SalesForecast struct {
 	Date       string  `json:"date"`
-	Predicted  float64 `json:"predicted"`
-	LowerBound float64 `json:"lowerBound"`
-	UpperBound float64 `json:"upperBound"`
+	Predicted  int64  `json:"predicted"`
+	LowerBound  int64  `json:"lowerBound"`
+	UpperBound  int64  `json:"upperBound"`
 }
 
 type ProfitAnalysis struct {
 	Period  string  `json:"period"`
-	Revenue float64 `json:"revenue"`
-	Cost    float64 `json:"cost"`
-	Profit  float64 `json:"profit"`
+	Revenue  int64  `json:"revenue"`
+	Cost  int64  `json:"cost"`
+	Profit  int64  `json:"profit"`
 	Margin  float64 `json:"margin"`
 	Growth  float64 `json:"growth"`
 }
@@ -39,8 +39,8 @@ type DemandForecast struct {
 type AnomalyDetection struct {
 	Metric      string  `json:"metric"`
 	Date        string  `json:"date"`
-	Expected    float64 `json:"expected"`
-	Actual      float64 `json:"actual"`
+	Expected  int64  `json:"expected"`
+	Actual  int64  `json:"actual"`
 	Deviation   float64 `json:"deviation"`
 	IsAnomaly   bool    `json:"isAnomaly"`
 	Description string  `json:"description"`
