+++ internal/domain/commission.go
@@ -6,10 +6,10 @@ type CommissionRule struct {
 	ID         string    `gorm:"primaryKey" json:"id"`
 	Name       string    `json:"name"`
 	Type       string    `json:"type"`
-	Value      float64   `json:"value"`
+	Value  int64  `json:"value"`
 	TargetType string    `json:"targetType"`
 	TargetID   string    `json:"targetId,omitempty"`
-	MinAmount  float64   `json:"minAmount"`
+	MinAmount  int64  `json:"minAmount"`
 	IsActive   bool      `gorm:"default:true" json:"isActive"`
 	CreatedAt  time.Time `json:"createdAt"`
 	UpdatedAt  time.Time `json:"updatedAt"`
@@ -21,8 +21,8 @@ type CommissionPayment struct {
 	StaffName   string     `json:"staffName"`
 	SaleID      string     `json:"saleId,omitempty"`
 	RuleID      string     `json:"ruleId"`
-	Amount      float64    `json:"amount"`
-	BaseAmount  float64    `json:"baseAmount"`
+	Amount  int64  `json:"amount"`
+	BaseAmount  int64  `json:"baseAmount"`
 	PeriodStart string     `json:"periodStart"`
 	PeriodEnd   string     `json:"periodEnd"`
 	Status      string     `json:"status"`
@@ -33,12 +33,12 @@ type CommissionPayment struct {
 type StaffPerformance struct {
 	StaffID      string  `json:"staffId"`
 	StaffName    string  `json:"staffName"`
-	TotalSales   float64 `json:"totalSales"`
+	TotalSales  int64  `json:"totalSales"`
 	SalesCount   int     `json:"salesCount"`
-	AvgSaleValue float64 `json:"avgSaleValue"`
-	TotalReturns float64 `json:"totalReturns"`
+	AvgSaleValue  int64  `json:"avgSaleValue"`
+	TotalReturns  int64  `json:"totalReturns"`
 	ReturnsCount int     `json:"returnsCount"`
-	Commission   float64 `json:"commission"`
+	Commission  int64  `json:"commission"`
 	PeriodStart  string  `json:"periodStart"`
 	PeriodEnd    string  `json:"periodEnd"`
 }
