+++ internal/domain/finance.go
@@ -6,7 +6,7 @@ import "time"
 type Expense struct {
 	ID        string    `gorm:"primaryKey" json:"id"`
 	Title     string    `json:"title"`
-	Amount    float64   `json:"amount"`
+	Amount  int64  `json:"amount"`
 	Date      string    `json:"date"`
 	Category  string    `json:"category"`
 	Notes     string    `json:"notes,omitempty"`
@@ -19,7 +19,7 @@ type Discount struct {
 	ID       string  `gorm:"primaryKey" json:"id"`
 	Name     string  `json:"name"`
 	Type     string  `json:"type"` // percentage, fixed
-	Value    float64 `json:"value"`
+	Value  int64  `json:"value"`
 	MinQty   float64 `json:"minQty"`
 	IsActive bool    `json:"isActive"`
 }
@@ -31,8 +31,8 @@ type Shift struct {
 	StaffName string    `json:"staffName"`
 	StartTime int64     `json:"startTime"`
 	EndTime   int64     `json:"endTime,omitempty"`
-	StartCash float64   `json:"startCash"`
-	EndCash   float64   `json:"endCash,omitempty"`
+	StartCash  int64  `json:"startCash"`
+	EndCash  int64  `json:"endCash,omitempty"`
 	Status    string    `json:"status"` // active, closed
 	CreatedAt time.Time `json:"createdAt"`
 }
@@ -42,7 +42,7 @@ type CashMovement struct {
 	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
 	ShiftID   string    `gorm:"index" json:"shiftId"`
 	Type      string    `json:"type"` // in, out
-	Amount    float64   `json:"amount"`
+	Amount  int64  `json:"amount"`
 	Reason    string    `json:"reason"`
 	StaffID   string    `json:"staffId"`
 	Timestamp int64     `json:"timestamp"`
@@ -55,7 +55,7 @@ type PurchaseOrder struct {
 	SupplierID   string              `gorm:"index" json:"supplierId"`
 	SupplierName string              `json:"supplierName"`
 	Date         string              `json:"date"`
-	Total        float64             `json:"total"`
+	Total  int64  `json:"total"`
 	Status       string              `json:"status"` // pending, received, cancelled
 	Items        []PurchaseOrderItem `gorm:"foreignKey:OrderID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;" json:"items"`
 	CreatedAt    time.Time           `json:"createdAt"`
@@ -69,6 +69,6 @@ type PurchaseOrderItem struct {
 	ProductID string  `json:"productId"`
 	Name      string  `json:"name"`
 	Qty       float64 `json:"qty"`
-	Cost      float64 `json:"cost"`
-	Total     float64 `json:"total"`
+	Cost  int64  `json:"cost"`
+	Total  int64  `json:"total"`
 }
