+++ internal/domain/recurring.go
@@ -10,10 +10,10 @@ type RecurringInvoice struct {
 	StartDate     string                 `json:"startDate"`
 	EndDate       string                 `json:"endDate,omitempty"`
 	NextRunDate   string                 `gorm:"index" json:"nextRunDate"`
-	Subtotal      float64                `json:"subtotal"`
-	Discount      float64                `json:"discount"`
-	VAT           float64                `json:"vat"`
-	Total         float64                `json:"total"`
+	Subtotal  int64  `json:"subtotal"`
+	Discount  int64  `json:"discount"`
+	VAT  int64  `json:"vat"`
+	Total  int64  `json:"total"`
 	PaymentMethod string                 `json:"paymentMethod"`
 	Status        string                 `gorm:"index" json:"status"`
 	Items         []RecurringInvoiceItem `gorm:"foreignKey:InvoiceID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;" json:"items"`
@@ -28,7 +28,7 @@ type RecurringInvoiceItem struct {
 	InvoiceID string  `gorm:"index" json:"-"`
 	ProductID string  `json:"productId"`
 	Name      string  `json:"name"`
-	Price     float64 `json:"price"`
+	Price  int64  `json:"price"`
 	Qty       float64 `json:"qty"`
-	Total     float64 `json:"total"`
+	Total  int64  `json:"total"`
 }
