+++ internal/domain/wallet.go
@@ -5,8 +5,8 @@ import "time"
 type CustomerWallet struct {
 	ID               string    `gorm:"primaryKey" json:"id"`
 	CustomerID       string    `gorm:"uniqueIndex" json:"customerId"`
-	Balance          float64   `json:"balance"`
-	CreditLimit      float64   `json:"creditLimit"`
+	Balance  int64  `json:"balance"`
+	CreditLimit  int64  `json:"creditLimit"`
 	AutoDebitEnabled bool      `gorm:"default:false" json:"autoDebitEnabled"`
 	AutoDebitDay     int       `json:"autoDebitDay,omitempty"`
 	CreatedAt        time.Time `json:"createdAt"`
@@ -16,7 +16,7 @@ type CustomerWallet struct {
 type WalletTransaction struct {
 	ID          uint      `gorm:"primaryKey;autoIncrement" json:"id"`
 	CustomerID  string    `gorm:"index" json:"customerId"`
-	Amount      float64   `json:"amount"`
+	Amount  int64  `json:"amount"`
 	Type        string    `json:"type"`
 	ReferenceID string    `json:"referenceId,omitempty"`
 	Description string    `json:"description"`
